Rename `extend` to `utilsExtend`
Running a project's CI with `phantomjs@~1.9.0` raises the following
error:

```
Exception while running qunit-bdd `before` hook:
`undefined` is not a function
  (evaluating `extend['default'].bind(Model, null)`)
```

`extend` seems to be a pre-existing variable (and will probably be a
keyword in future browsers). Therefore `extend['default']` is
`undefined`.

To work around this, export the `utils/extend` to a different name.