[#607] Drop Collection.[], add Collection.models
Collection[n] was a confusing API because it's not a true array, but having
the bracket accessors suggested it was. This lead to confusing situations
where collections wouldn't work as expected with e.g. Lodash methods.

The [] syntax no longer works, and collections now expose their underlying
models on a `.models` property that is a plain JavaScript array. If users need
the full array they can use this with any library they want.