allow models to be updated by mirage created via a factory
We were seeing consistently a case where creating a new model, then
updating via the UI wouldn't persist changes to mirage's db.

I tracked this down to `this.save()` being called with `this.attrs`.
There were missing Object descriptors on the model due to mirage
skipping properties that were not undefined, even though there were
no descriptors for them.

To fix this, I used Object.getDescriptor() to introspect on the model
to see if the property has been defined. This fixes the issue and allows
for updates to pass through smoothly.