Fix JSONAPISerializer#_getResourceObjectForModel method
The method fixed according to JSONAPI spec. The spec says that a "relationships" object must contain at least one of the props, but could contain all of them. So, `links` are able to be with `data` as well.
Link to the article of the spec: http://jsonapi.org/format/#document-resource-object-relationships

fixes #973