Base Class.extend() on native es class syntax
The extend mechanism, which we inherited from Ember pre-dates the
advent of native ES class syntax. Among other things, this meant
manually constructing the prototype chain when expressing the
inheritance.

This change lets the `class` syntax do the heavy lifting of prototype
wiring while making the code more clear as to its intent.

As a bonus, instances created with this class show up in development
tools as `class Child` of instead of

```
function () {
  return parent.apply(this, arguments);
}
```

Still not ideal, but better than the alternative.

Finally, the `__super__` property was removed since it's of dubious
value and did not cause any tests to fail.