feat: `normalizeIds` serializer property defaults to true
BREAKING CHANGE: This applies to the ActiveModelSerializer and
RestSerializer.

The `normalize` property on serializers helps Mirage's shorthands work
by transforming differently formatted payloads into JSON:API documents.
These documents are then used by Mirage to update the database accordingly.

There was a gap in the default `normalize` method for a long time, in that
it didn't take REST payloads that included foreign keys and looked like

```js
let payload = {
  contact: {
    id: 1,
    name: 'Link',
    address: 1
  }
};
```

and turn that `address` key into a proper JSON:API relationship:

```js
data: {
  type: 'contacts',
  id: 1,
  attributes: {
    name: 'Link'
  },
  relationships: {
    address: {
      data: {
        type: 'address',
        id: 1
      }
    }
  }
}
```

We added this feature a while ago, and it's controlled with the `normalizeIds`
property on the ActiveModelSerializer and RESTSerializer. (We did this
so the feature wouldn't be a breaking change.)

We're now making `true` the default, which should be the behavior
everyone desires (assuming they're using shorthands). This is technically
a breaking change, though it's unlikely to affect most people.