fix: updating model w/two same-model relationships
For a model that has two inverse one-to-many relationships that point
to the same model class, there was a bug where calling update on both
at the same time

```js
post.update({
  author: user,
  editor: user
});
```

would fail, whereas updating each relationship individually would work.
This is because the temporary model reference for the second
relationship became stale after the first one was saved.

See #1613.