Add Fastboot to addon blacklist for dummy app
We have ember-cli-fastboot installed so it can be made available to some
of the test projects, but we never use it to actually serve the dummy
app (which is the docs app).

Thus, if a new user clones + runs the repo, unless they have the
FASTBOOT_DISABLED ENV var set, `ember s` will try to serve the
dummy app via FastBoot which will cause errors.

With the addon blacklisted, ember-cli-fastboot will just never be able
to affect the dummy app at all.