Fix invalid use of ES modules
The ES module spec doesn't support "optional" exports. If you try import a name from a module and that name is not exported, it's a SyntaxError.

Mirage tries to import `default`, `testConfig`, and `makeServer` from the user's config, but they are treated as optional.

The solution is to import the module's whole namespace (`import * as...`) and then inspect the namespace object.